feat: Add Juri and Cammy combo pages (#22)
This commit adds two new pages, "Juri" and "Cammy", to the application. Each page displays a combo list for the respective character in Street Fighter 6. The combo data is loaded from JSON files and rendered using the Svgloader component. The pages also include a title and a link to the GitHub profile of the creator.